+++ sidecar/run-emitter.js
@@ -1,0 +1,3 @@
+'use strict';
+function normalizeRunEmitter(value) { return typeof value === 'function' ? value : function () {}; }
+module.exports = { normalizeRunEmitter };